# docs/seo-brain/openapi.v1.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SEO Brain API",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/v1/sites": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "List Sites",
    "operationId": "list_sites_api_v1_sites_get",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": true
         },
         "title": "Response List Sites Api V1 Sites Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "sites"
    ],
    "summary": "Create Site",
    "operationId": "create_site_api_v1_sites_post",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Site Api V1 Sites Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Get Site",
    "operationId": "get_site_api_v1_sites__site_id__get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Site Api V1 Sites  Site Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "sites"
    ],
    "summary": "Update Site",
    "operationId": "update_site_api_v1_sites__site_id__patch",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update Site Api V1 Sites  Site Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "sites"
    ],
    "summary": "Delete Site",
    "description": "Delete a site. Refuses (409 `site_has_data`) when related rows exist unless force=true.\nFiles in the workspace directory are never deleted by the API.",
    "operationId": "delete_site_api_v1_sites__site_id__delete",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "force",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "description": "also delete all site data (graph, GSC, crawl, memory)",
       "default": false,
       "title": "Force"
      },
      "description": "also delete all site data (graph, GSC, crawl, memory)"
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Delete Site Api V1 Sites  Site Id  Delete"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/connections": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Get Connections",
    "description": "Last known status per connection kind (gsc | ga4 | wordpress) + what is configured on the site.",
    "operationId": "get_connections_api_v1_sites__site_id__connections_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Connections Api V1 Sites  Site Id  Connections Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/connections/{kind}/test": {
   "post": {
    "tags": [
     "sites"
    ],
    "summary": "Test Connection",
    "description": "Run a read-only permission test. If `property` is given and the test passes, it is stored on the site.",
    "operationId": "test_connection_api_v1_sites__site_id__connections__kind__test_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Kind"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/ConnectionTestRequest"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Test Connection Api V1 Sites  Site Id  Connections  Kind  Test Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sites/{site_id}/initialize": {
   "post": {
    "tags": [
     "sites"
    ],
    "summary": "Initialize Site",
    "description": "Wizard step 3: workspace + site memory + graph namespace (idempotent).",
    "operationId": "initialize_site_api_v1_sites__site_id__initialize_post",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Initialize Site Api V1 Sites  Site Id  Initialize Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/connections/gsc/properties": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Gsc Properties",
    "description": "Properties visible to the connected Google account (for the wizard's dropdown).",
    "operationId": "gsc_properties_api_v1_connections_gsc_properties_get",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Gsc Properties Api V1 Connections Gsc Properties Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/ai/routes": {
   "get": {
    "tags": [
     "ai"
    ],
    "summary": "Routes",
    "operationId": "routes_api_v1_ai_routes_get",
    "parameters": [
     {
      "name": "x-api-token",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Routes Api V1 Ai Routes Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ConnectionTestRequest": {
    "properties": {
     "property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Property"
     }
    },
    "type": "object",
    "title": "ConnectionTestRequest",
    "description": "Optional override of the property to test; when omitted the site's stored property is used."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "SiteCreate": {
    "properties": {
     "site_id": {
      "type": "string",
      "pattern": "^[a-z0-9][a-z0-9\\-]{1,62}$",
      "title": "Site Id",
      "description": "stable slug, e.g. emdadmodiran"
     },
     "timezone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timezone",
      "default": "Asia/Tehran"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "canonical_url": {
      "type": "string",
      "maxLength": 2083,
      "minLength": 1,
      "format": "uri",
      "title": "Canonical Url"
     },
     "wp_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2083,
        "minLength": 1,
        "format": "uri"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wp Url"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language",
      "default": "fa-IR"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country",
      "default": "IR"
     },
     "business_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Business Type"
     },
     "gsc_property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gsc Property"
     },
     "ga4_property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ga4 Property"
     },
     "mode": {
      "type": "string",
      "enum": [
       "manual",
       "assisted",
       "autopilot"
      ],
      "title": "Mode",
      "default": "manual"
     }
    },
    "type": "object",
    "required": [
     "site_id",
     "name",
     "canonical_url"
    ],
    "title": "SiteCreate"
   },
   "SiteUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "timezone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timezone"
     },
     "wp_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2083,
        "minLength": 1,
        "format": "uri"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wp Url"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language"
     },
     "country": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Country"
     },
     "business_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Business Type"
     },
     "gsc_property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Gsc Property"
     },
     "ga4_property": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ga4 Property"
     },
     "mode": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "manual",
         "assisted",
         "autopilot"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Mode"
     }
    },
    "type": "object",
    "title": "SiteUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}